#JENIOS-15 Add first tests for Builds.
