+++ JenkinsiOSTests/Build Tests/BuildTestsMinimalResource.json
@@ -1,0 +1,5 @@
+{
+    "_class": "hudson.maven.MavenModuleSetBuild",
+    "number": 122,
+    "url": "https://builds.apache.org/job/Accumulo-1.8/122/"
+}
